pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swagger-pet-tester"
version = "1.0.0"
description = "API autotests for Petstore Swagger"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["api-testing", "automation", "petstore", "pytest", "rest-api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "requests>=2.31.0",
    "pytest>=8.0.0",
    "pytest-html>=4.1.1",
    "faker>=22.0.0",
    "pydantic>=2.5.3",
    "allure-pytest>=2.13.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "ruff>=0.1.0",
]
report = [
    "allure-pytest>=2.13.2",
    "pytest-html>=4.1.1",
    "pytest-json-report>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/your-username/swagger-pet-tester"
Repository = "https://github.com/your-username/swagger-pet-tester"
Issues = "https://github.com/your-username/swagger-pet-tester/issues"

# ============================================
# Tool configurations
# ============================================

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | venv
  | env
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
exclude = [
    "venv/",
    "env/",
    "build/",
    "dist/",
    "reports/",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
    "-p", "no:cacheprovider",
]
markers = [
    "smoke: Быстрая проверка основного функционала",
    "regression: Полный набор регрессионных тестов",
    "negative: Негативные тесты (ошибки, граничные случаи)",
    "pet: Тесты для Pet эндпоинтов",
    "store: Тесты для Store эндпоинтов",
    "user: Тесты для User эндпоинтов",
    "slow: Медленные тесты (загрузка файлов и т.д.)",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%H:%M:%S"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
timeout = 30

[tool.coverage.run]
source = ["api", "utils", "models"]
omit = [
    "tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
]
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    "env",
    "build",
    "dist",
    "reports",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]  # assert used in tests

[tool.ruff.isort]
known-first-party = ["api", "config", "models", "tests", "utils"]